# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "鸵小妥API",
        "contact": {
            "email": "wq0426@163.com"
        },
        "version": "1.0.2"
    },
    "host": "127.0.0.1:8289",
    "paths": {
...
        "/order/create": {
            "post": {
                "description": "创建用户订单，支持多个商品一次性提交",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "创建用户订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建订单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/order/detail/{order_item_id}": {
            "get": {
                "description": "获取订单详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrderDetailResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "404": {
                        "description": "订单不存在",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/order/list": {
            "get": {
                "description": "获取用户订单列表，可根据状态筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取订单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "订单状态：0-待付款，1-已付款，2-已发货，3-已完成，4-已取消；不传则查询所有状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认10",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrderListResponse"
                        }
                    }
                }
            }
        },
        "/order/products": {
            "get": {
                "description": "根据订单ID查询该订单下的所有商品和优惠券信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取订单商品详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "订单ID",
                        "name": "order_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "订单商品详情",
                        "schema": {
                            "$ref": "#/definitions/model.OrderProductsResponse"
                        }
                    }
                }
            }
        },
        "/order/status": {
            "post": {
                "description": "根据订单ID更新订单状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "更新订单状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "更新订单状态请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误或非法的状态转换",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "404": {
                        "description": "订单不存在或无权限",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isDefault": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                }
            }
        },
        "model.AddressInfo": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "城市",
                    "type": "string"
                },
                "detail_address": {
                    "description": "详细地址",
                    "type": "string"
                },
                "district": {
                    "description": "区县",
                    "type": "string"
                },
                "province": {
                    "description": "省份",
                    "type": "string"
                },
                "receiver_name": {
                    "description": "收货人姓名",
                    "type": "string"
                },
                "receiver_phone": {
                    "description": "收货人电话",
                    "type": "string"
                }
            }
        },
...
        "model.CreateOrderRequest": {
            "type": "object",
            "required": [
                "address",
                "items",
                "payment_method"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/model.Address"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrderItemRequest"
                    }
                },
                "payment_method": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.OrderDetailResponse": {
            "type": "object",
            "properties": {
                "actual_amount": {
                    "description": "实际支付金额",
                    "type": "number"
                },
                "address": {
                    "description": "地址信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AddressInfo"
                        }
                    ]
                },
                "completed_at": {
                    "description": "完成时间",
                    "type": "string"
                },
                "coupon_discount": {
                    "description": "优惠券折扣",
                    "type": "number"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "member_discount": {
                    "description": "会员折扣",
                    "type": "number"
                },
                "order_amount": {
                    "description": "订单金额",
                    "type": "number"
                },
                "order_id": {
                    "description": "订单基本信息",
                    "type": "integer"
                },
                "order_no": {
                    "description": "订单号",
                    "type": "string"
                },
                "pay_method": {
                    "description": "支付方式",
                    "type": "integer"
                },
                "pay_method_text": {
                    "description": "支付方式文本",
                    "type": "string"
                },
                "pay_time": {
                    "description": "支付时间",
                    "type": "string"
                },
                "products": {
                    "description": "商品信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductListItemDTO"
                    }
                },
                "shipped_at": {
                    "description": "发货时间",
                    "type": "string"
                },
                "shipping_fee": {
                    "description": "运费",
                    "type": "number"
                },
                "status": {
                    "description": "订单状态",
                    "type": "integer"
                },
                "status_text": {
                    "description": "订单状态文本",
                    "type": "string"
                },
                "store_id": {
                    "description": "店铺ID",
                    "type": "integer"
                },
                "store_logo": {
                    "description": "店铺Logo",
                    "type": "string"
                },
                "store_name": {
                    "description": "店铺名称",
                    "type": "string"
                },
                "total_price": {
                    "description": "订单汇总信息",
                    "type": "number"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                }
            }
        },
...
        "model.OrderItemRequest": {
            "type": "object",
            "required": [
                "current_price",
                "image",
                "product_id",
                "product_name",
                "quantity"
            ],
            "properties": {
                "cart_id": {
                    "type": "integer"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_price": {
                    "type": "number"
                },
                "courier_fee_min": {
                    "type": "number"
                },
                "current_price": {
                    "type": "number"
                },
                "image": {
                    "type": "string"
                },
                "member_discount": {
                    "type": "number"
                },
                "note": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "store_id": {
                    "type": "integer"
                },
                "store_name": {
                    "type": "string"
                }
            }
        },
        "model.OrderListDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_no": {
                    "type": "string"
                },
                "pay_time": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/model.OrderProductDTO"
                },
                "status": {
                    "type": "integer"
                },
                "status_text": {
                    "type": "string"
                },
                "store_icon": {
                    "description": "店铺图标",
                    "type": "string"
                },
                "store_name": {
                    "description": "店铺名称",
                    "type": "string"
                },
                "total_fee": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.OrderListResponse": {
            "type": "object",
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrderListDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.OrderProductDTO": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "item_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "model.OrderProductsResponse": {
            "type": "object",
            "properties": {
                "products": {
                    "description": "商品列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductListItemDTO"
                    }
                }
            }
        },
...
        "model.ProductCouponDTO": {
            "type": "object",
            "properties": {
                "available_min_price": {
                    "type": "number"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_name": {
                    "type": "string"
                },
                "coupon_price": {
                    "type": "number"
                },
                "deadline": {
                    "type": "string"
                },
                "is_received": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "model.ProductEvaluateDTO": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "evaluate_content": {
                    "type": "string"
                },
                "evaluate_id": {
                    "type": "integer"
                },
                "evaluate_images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "evaluate_nums": {
                    "type": "integer"
                },
                "evaluate_time": {
                    "type": "string"
                },
                "is_return": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "praise_nums": {
                    "type": "integer"
                },
                "view_nums": {
                    "type": "integer"
                }
            }
        },
...
        "model.ProductListItemDTO": {
            "type": "object",
            "properties": {
                "category1_id": {
                    "type": "integer"
                },
                "category2_id": {
                    "type": "integer"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_price": {
                    "type": "number"
                },
                "courier_fee_max": {
                    "type": "number"
                },
                "courier_fee_min": {
                    "type": "number"
                },
                "header_img": {
                    "type": "string"
                },
                "item_id": {
                    "type": "integer"
                },
                "member_discount": {
                    "type": "number"
                },
                "product_content": {
                    "type": "string"
                },
                "product_coupons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductCouponDTO"
                    }
                },
                "product_current_price": {
                    "type": "number"
                },
                "product_evaluate": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductEvaluateDTO"
                    }
                },
                "product_evaluate_nums": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "product_is_special": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "product_original_price": {
                    "type": "number"
                },
                "product_quantity": {
                    "type": "integer"
                },
                "product_sales": {
                    "type": "integer"
                },
                "product_spec": {
                    "type": "string"
                },
                "product_specification": {
                    "type": "string"
                },
                "product_unit": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateOrderStatusRequest": {
            "type": "object",
            "required": [
                "order_id"
            ],
            "properties": {
                "order_id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "status": {
                    "description": "订单状态",
                    "type": "integer"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}